feat!: remove unnecessary `sched.Schedule()` func
`Schedule()` function used to be called on Scheduler and returned `*Job`
that was ready to be filled with call chain starting with `Every()`.
This ment that `Schedule()` had to be called (without arguments!) every
time a new task was being scheduled. In order to remove this boiler
code, `Every()` and `EverySingle()` functions are now defined on the
`*Scheduler` and create `*Job` before continuing their usual work.

This is a breaking change to the API.